pred_t is long int now; introduced MAX_PRED.
The only practical advantage over int is that it has the same size as
void* and we can thus use it with our containers.

Like MAX_STD_NAME, BATs define a MAX_PRED now.

+++ src/literal.h
@@ -14,7 +14,7 @@
 
 #include "term.h"
 
-typedef int pred_t;
+typedef long int pred_t;
 typedef struct {
     stdvec_t z;
     bool sign;
